requirements.txt
# Security Incident Detection & Response System - Requirements

# Backend API
flask>=2.0.0
sqlalchemy>=1.4.0

# Dashboard
streamlit>=1.28.0
pandas>=1.5.0
fpdf2>=2.7.0

# Detection Engine
requests>=2.28.0
psutil>=5.9.0

# Data Generation (optional)
faker>=18.0.0

